bonbon v0.0.1 Bonbon.Model.Locale

A model representing the different languages using culture codes (ISO 3166-1 alpha-2 and ISO 639-1 code).

Functions

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

to_locale_id(arg)

Specs

to_locale_id(String.t) :: integer | nil

Get the locale_id for the given string or nil on an invalid locale.

The string format takes the form of language_country or language when no country is specified. e.g. "en" and "en_AU" would be valid formats, the first referring to the english locale, the second referring to Australian english.

to_locale_id!(code)

Specs

to_locale_id!(String.t) :: integer

to_locale_id!(String.t) :: [integer]

Get the locale_id for the given string or raise the exception Bonbon.Model.Locale.NotFoundError on an invalid locale. For more details see: to_locale_id/1.

to_locale_id_list(arg)

Specs

to_locale_id_list(String.t) :: [integer] | nil

Get the fallback list of locale_id's for the given string or empty list if no locales were valid.

The string format takes the form of language_country or language when no country is specified. e.g. "en" and "en_AU" would be valid formats, the first referring to the english locale, the second referring to Australian english.

This list includes the top-most locale, and parent locales (to fallback to).

to_locale_id_list!(code)

Get the fallback list of locale_id's for the given string or raise the exception Bonbon.Model.Locale.NotFoundError when no locales are found. For more details see: to_locale_id_list/1.